[update] added Authentication and resources

+++ app/Http/Controllers/CarController.php
@@ -4,24 +4,27 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\CarRequest;
 use App\Models\Car;
-use App\Models\User;
-use Illuminate\Http\Request;
 
 class CarController extends Controller
 {
     public function index()
     {
-        return response()->json(['cars' => Car::all()]);
+        return response()->json([
+            'cars' => auth()->user()->cars
+        ]);
     }
 
     public function show(Car $car)
     {
-        return response()->json(['car' => $car]);
+        $this->authorize('view', $car);
+        return response()->json([
+            'car' => $car
+        ]);
     }
 
     public function store(CarRequest $request)
     {
-        $car = User::first()->cars()->create($request->validated());
+        auth()->user()->cars()->create($request->validated());
         return response()->json([
             'msg' => 'ok',
             'car' => $car
@@ -30,6 +33,7 @@ class CarController extends Controller
 
     public function update(Car $car, CarRequest $request)
     {
+        $this->authorize('update', $car);
         $car->update($request->validated());
         return response()->json([
             'msg' => 'ok',
@@ -39,6 +43,7 @@ class CarController extends Controller
 
     public function destroy(Car $car)
     {
+        $this->authorize('delete', $car);
         $car->delete();
         return response()->json([
             'msg' => 'ok'
